updated controllers
